+++ CHANGELOG.md
@@ -91,7 +91,7 @@
 - [4826](https://github.com/vegaprotocol/vega/issues/4826) - Tidying up feature tests in verified folder
 - [4843](https://github.com/vegaprotocol/vega/issues/4843) - Make snapshot engine aware of local storage old versions, and manage them accordingly to stop growing disk usage.
 - [4863](https://github.com/vegaprotocol/vega/issues/4863) - Improve replay protection
-
+- [4867](https://github.com/vegaprotocol/vega/issues/4867) - Optimise replay protection
 
 ### 🐛 Fixes
 - [4842](https://github.com/vegaprotocol/vega/pull/4842) - Fix margin balance not being released after close-out.